Add initial implementation of json-log-parser
This is a command line application that can be used to parse Java json logs and output them as formatted plain text.

Structured logging using json as the format is very popular in cloud scenarios as it integrates with cloud tooling seamlessly. However, these logs are harder to read, with stacktraces being especially hard to read.

This is a simple tool that can be used to aid debugging and development work.

The tool, by default, takes input from stdin and writes output to stdout.

See README.md for further details.